+++ src/de/error.rs
@@ -1044,6 +1044,12 @@ pub enum Error {
         source: ValidationSource,
         errors: Vec<Error>,
     },
+
+    /// An explicitly specified YAML tag is not supported by this crate.
+    UnsupportedTag {
+        tag: String,
+        location: Location,
+    },
 }
 
 impl Error {
@@ -1398,6 +1404,7 @@ impl Error {
             | Error::Unexpected { location, .. }
             | Error::MergeValueNotMapOrSeqOfMaps { location }
             | Error::MergeKeyNotAllowed { location }
+            | Error::UnsupportedTag { location, .. }
             | Error::InvalidBinaryBase64 { location }
             | Error::BinaryNotUtf8 { location }
             | Error::TaggedScalarCannotDeserializeIntoString { location }
@@ -1511,6 +1518,7 @@ impl Error {
             | Error::Unexpected { location, .. }
             | Error::MergeValueNotMapOrSeqOfMaps { location }
             | Error::MergeKeyNotAllowed { location }
+            | Error::UnsupportedTag { location, .. }
             | Error::InvalidBinaryBase64 { location }
             | Error::BinaryNotUtf8 { location }
             | Error::TaggedScalarCannotDeserializeIntoString { location }